+++ app/src/main/java/es/situm/gettingstarted/common/floorselector/FloorAdapter.java
@@ -1,32 +1,36 @@
-package es.situm.gettingstarted.drawbuilding;
+package es.situm.gettingstarted.common.floorselector;
 
+import android.content.Context;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
-
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.recyclerview.widget.RecyclerView;
-
 import java.util.List;
-
 import es.situm.gettingstarted.R;
 import es.situm.sdk.model.cartography.Floor;
 
 public class FloorAdapter extends RecyclerView.Adapter<FloorAdapter.ViewHolder> {
 
     private final OnItemClickListener onItemClickListener;
-
-    @Nullable
+    private final Context context;
     private Floor selected;
+    private Floor positioningFloor;
     private List<Floor> floors;
 
-    FloorAdapter(List<Floor> floors, OnItemClickListener onItemClickListener) {
-        this.onItemClickListener = onItemClickListener;
+    FloorAdapter(List<Floor> floors, OnItemClickListener onItemClickListener, Context context) {
         setFloors(floors);
+        this.onItemClickListener = onItemClickListener;
+        this.context = context;
     }
 
+    /**
+     * Set the floors of the building
+     *
+     * @param floors List<Floor>
+     */
     void setFloors(List<Floor> floors) {
         this.floors = floors;
         notifyItemRangeChanged(0, getItemCount());
@@ -56,12 +60,33 @@ public class FloorAdapter extends RecyclerView.Adapter<FloorAdapter.ViewHolder> 
         return selected;
     }
 
-    public void select(Floor floor) {
+    /**
+     * Select the floor provided
+     *
+     * @param floor Floor
+     */
+    void select(Floor floor) {
         int indexPrevSelected = floors.indexOf(selected);
         int indexNewSelected = floors.indexOf(floor);
         selected = floor;
         notifyItemChanged(indexPrevSelected);
         notifyItemChanged(indexNewSelected);
+    }
+
+    /**
+     * Updates the positioning floor
+     *
+     * @param newPositioningFloor Floor
+     */
+    void positioningFloorChangedTo(Floor newPositioningFloor, boolean userPickedFloor) {
+        int indexPrevPositioned = floors.indexOf(positioningFloor);
+        int indexNewPosition = floors.indexOf(newPositioningFloor);
+        if(userPickedFloor){
+            select(newPositioningFloor);
+        }
+        positioningFloor = newPositioningFloor;
+        notifyItemChanged(indexPrevPositioned);
+        notifyItemChanged(indexNewPosition);
     }
 
     class ViewHolder extends RecyclerView.ViewHolder {
@@ -73,12 +98,16 @@ public class FloorAdapter extends RecyclerView.Adapter<FloorAdapter.ViewHolder> 
         }
 
         void bind(Floor floor) {
+            //By default the item is not SELECTED or POSITIONING
             tvLevelNumber.setText(FloorUtils.getNameOrLevel(floor));
+            itemView.setBackgroundResource(R.drawable.situm_item_level_background);
+            tvLevelNumber.setTextAppearance(context, R.style.situm_normalText);
 
-            if (selected != null && selected.getIdentifier().equals(floor.getIdentifier())) {
+            if(positioningFloor != null && positioningFloor.equals(floor)){ // POSITIONING
+                itemView.setBackgroundResource(R.drawable.situm_item_level_background_positioning);
+                tvLevelNumber.setTextAppearance(context, R.style.situm_boldText);
+            }else if (selected != null && selected.equals(floor)){ // SELECTED
                 itemView.setBackgroundResource(R.drawable.situm_item_level_background_selected);
-            } else {
-                itemView.setBackgroundResource(R.drawable.situm_item_level_background);
             }
 
             itemView.setOnClickListener(v -> onItemClickListener.onItemClick(floor));