+++ CHANGELOG.md
@@ -9,7 +9,19 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ### Added
 
-- Added floor selector in Draw building over map example
+- Added new permission on AndroidManifest.xml (BLUETOOTH_SCAN) for Android 12 devices that need to request this permission (Pixel 4a for example)
+
+### Fixed
+
+- Fixed the strange behavior of the selector right after leaving positioning mode.
+
+### Modified
+
+- Abstraction of the floor selector & implementation of it in 'Animate position while walking' activity
+
+### Added
+
+- Added floor selector in 'Draw building over map' example
 
 ## [0.0.1] - 2022-02-24